Move Property into property-descriptor.ts
